+++ eox_core/api/data/aggregated_collector/queries.py
@@ -14,6 +14,7 @@ SELECT
 FROM
   auth_user as au;
 """
+
 # This query counts the number of unique active users in the last month.
 # A user is considered active if they have modified a student module within the last month.
 ACTIVE_USERS_LAST_MONTH_QUERY = """
@@ -29,6 +30,7 @@ WHERE
 GROUP BY
     YEAR(cs.modified), MONTH(cs.modified);
 """
+
 # This query counts the number of unique active users in the last 7 days.
 # A user is considered active if they have modified a student module within the last 7 days.
 ACTIVE_USERS_LAST_7_DAYS_QUERY = """
@@ -39,6 +41,7 @@ FROM
 WHERE
   cs.modified >= DATE_SUB(CURDATE(), INTERVAL 7 DAY);
 """
+
 # This query counts the total number of courses created on the platform.
 TOTAL_COURSES_CREATED_QUERY = """
 SELECT
@@ -46,6 +49,7 @@ SELECT
 FROM
   course_overviews_courseoverview as coc;
 """
+
 # This query counts the number of CourseOverviews objects that started before
 # now and have not yet ended
 ACTIVE_COURSES_COUNT_QUERY = """
@@ -60,6 +64,7 @@ WHERE
     OR coc.end IS NULL
   );
 """
+
 # This query counts the number of courses that have at least one issued certificate.
 COURSES_WITH_ACTIVE_CERTIFICATES_QUERY = """
 SELECT
@@ -71,6 +76,7 @@ JOIN
 ON
   coc.id = cg.course_id;
 """
+
 # This query counts the number of new enrollments in the last month.
 ENROLLMENTS_LAST_MONTH_QUERY = """
 SELECT
@@ -85,6 +91,7 @@ WHERE
 GROUP BY
   YEAR(sc.created), MONTH(sc.created);
 """
+
 # This query counts the number of new enrollments in the last 7 days.
 ENROLLMENTS_LAST_7_DAYS_QUERY = """
 SELECT
@@ -94,13 +101,15 @@ FROM
 WHERE
   sc.created >= DATE_SUB(CURDATE(), INTERVAL 7 DAY);
 """
+
+# This query counts the total number of certificates issued on the platform.
 CERTIFICATES_ISSUED_QUERY = """
 SELECT
   COUNT(*)
 FROM
   certificates_generatedcertificate as cg;
 """
-# This query counts the total number of certificates issued on the platform.
+
 PREDEFINED_QUERIES = {
     "Total Users": TOTAL_USERS_QUERY,
     "Active Users Last Month": ACTIVE_USERS_LAST_MONTH_QUERY,